refactor(backlinkGenerator): move backlinkGenerator to src directory
- Relocate backlinkGenerator.ts from root to src directory
- Maintain existing functionality without changes
- Update file path references accordingly
